+++ DOficiales.php
@@ -1,31 +1,43 @@
-<html>
-<!-- Solicitar a traves de un formulario el IDconductor -->
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Eliminar Oficial</title>
+</head>
+
+<body>
+    <!-- Solicitar a traves de un formulario el IDconductor -->
     <form method="GET" action="DOficiales.php">
         <label> ID Oficial </label>
-        <input type="text" name="IDOficial" id="IDOficial" placeholder="8409">
+        <input type="text" name="IDOficial" id="IDOficial">
         <input type="submit" value="Enviar">
     </form>
+</body>
+
 </html>
 
-<?php 
+<?php
 
 // Eliminar el registro de la tabla oficiales 
-if(isset($_GET['IDOficial'])){
+if (isset($_GET['IDOficial']) && !empty($_GET['IDOficial'])) {
 
     $IDOficial = $_GET['IDOficial'];
 
     $SQL = "DELETE FROM oficiales WHERE IDOficial = '$IDOficial'";
-    print("<br>".$SQL);
 
     //enviar al dbms
     include("conexion.php");
     $Con = Conectar();
-    $Result = Ejecutar($Con, $SQL) or die ("Error al eliminar datos".mysqli_error($Con));
-    $FilasAfectadas = mysqli_affected_rows($Con);
+    $Result = Ejecutar($Con, $SQL) or die("Error al eliminar datos" . mysqli_error($Con));
 
-    print("<br> Numero de Filas Eliminadas:" . $FilasAfectadas);
+    print("<br> Numero de Filas Eliminadas: " . mysqli_affected_rows($Con));
 
     Cerrar($Con);
+} else {
+    print("<br> No se ha recibido el ID del oficial");
 }
 
 ?>
