+++ DPropietarios.php
@@ -1,21 +1,28 @@
-<html>
-<!-- Solicitar a traves de un formulario el RFC -->
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Eliminar propietario</title>
+</head>
+<body>
     <form method="GET" action="DPropietarios.php">
         <label> RFC del propietario </label>
         <input type="text" name="RFC" id="RFC" placeholder="VECJ880326XXX">
         <input type="submit" value="Enviar">
     </form>
+</body>
 </html>
 
 <?php 
 
 // Eliminar el registro de la tabla oficiales 
-if(isset($_GET['RFC'])){
+if(isset($_GET['RFC']) && !empty($_GET['RFC'])){
 
     $RFC = $_GET['RFC'];
 
     $SQL = "DELETE FROM propietarios WHERE RFC = '$RFC'";
-    print("<br>".$SQL);
 
     //enviar al dbms
     include("conexion.php");
@@ -26,6 +33,8 @@ if(isset($_GET['RFC'])){
     print("<br> Numero de Filas Eliminadas:" . $FilasAfectadas);
 
     Cerrar($Con);
+} else {
+    print("No se ha recibido el RFC del propietario");
 }
 
 ?>
